修正

+++ db/migrate/20231101125738_create_group_tags.rb
@@ -1,8 +1,8 @@
 class CreateGroupTags < ActiveRecord::Migration[6.1]
   def change
     create_table :group_tags do |t|
-      t.references :group, type: :bigint(20), foreign_key: true
-      t.references :tag, type: :bigint(20), foreign_key: true
+      t.references :group, type: :integer, foreign_key: true,
+      t.references :tag, type: :bigint, foreign_key: true
 
       t.timestamps
     end
